Fix OpenAI API error by casting max_tokens to integer
- Explicitly cast max_tokens parameter to integer in API requests
- Fixes "Invalid type for 'max_tokens': expected an integer, but got a string instead" error

+++ wp-content/plugins/chatbot-plugin/includes/class-chatbot-openai.php
@@ -293,7 +293,7 @@ class Chatbot_OpenAI {
             $request_body = array(
                 'model' => $model,
                 'messages' => $messages,
-                'max_tokens' => $this->max_tokens,
+                'max_tokens' => (int) $this->max_tokens,
                 'temperature' => (float) $this->temperature,
             );
             
@@ -479,7 +479,7 @@ class Chatbot_OpenAI {
         $request_body = array(
             'model' => $model,
             'messages' => $test_message,
-            'max_tokens' => 10,
+            'max_tokens' => (int) 10,
             'temperature' => 0.1,
         );
